+++ service/internal/bootstrap/bootstrap.go
@@ -127,6 +127,16 @@ func (b *Bootstrap) Run(ctx context.Context) error {
 		return microerror.Mask(err)
 	}
 
+	err = b.createWriteNodePoolsClusterRole(ctx)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = b.createWriteNodePoolsClusterRoleBindingToAutomationSA(ctx)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
 	err = b.labelDefaultClusterRoles(ctx)
 	if err != nil {
 		return microerror.Mask(err)